refactor(hooks): extract targets.rs submodule from src/hooks/mod.rs (#2188)
Closes #2188
